requirements.txt
streamlit==1.28.1
langchain-groq==0.0.6
langchain==0.1.0
python-dotenv==1.0.0
sounddevice==0.4.6
scipy==1.11.4
openai-whisper==20231117
numpy==1.24.3
requests==2.31.0
elevenlabs>=2.8.0 